scripts/requirements.txt
questionary==1.9.0
unidecode==1.2.0
pillow==8.2.0
jinja2==3.0.1
pyyaml==5.4.1
python-slugify==5.0.2
python-docx==0.8.11